requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
newsapi-python==0.2.5
pylint==2.3.1
python-dotenv==0.10.3
requests==2.21.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.24.3
Werkzeug==0.15.5
wrapt==1.11.2
